Texture: Store dimensions in 16 bits
One can reasonably assume that the textures will not have dimensions of
more than 65536.

+++ src/Graphics/FrameBuffer.cpp
@@ -27,7 +27,7 @@
 
 using namespace openblack::graphics;
 
-FrameBuffer::FrameBuffer(uint32_t width, uint32_t height, Format colorFormat, std::optional<Format> depthStencilFormat) :
+FrameBuffer::FrameBuffer(uint16_t width, uint16_t height, Format colorFormat, std::optional<Format> depthStencilFormat) :
 	_handle(0), _width(width), _height(height), _colorFormat(colorFormat), _depthStencilFormat(depthStencilFormat)
 {
 	glGenFramebuffers(1, &_handle);
